Added frontend and moved broker auth to settings
