+++ app/templates/settings/_sidebar.html
@@ -10,7 +10,7 @@
         <hr>
         <b style="padding: 0.5rem 1rem;">Brokers</b>
         <li class="nav-item">
-            <a id="sb-pref" href="#" class="nav-link link-dark">Add Brokers</a>
+            <a id="sb-pref" href="{{ url_for('settings.brokers_settings') }}" class="nav-link link-dark">Add Brokers</a>
         </li>
         <hr>
         <b style="padding: 0.5rem 1rem;">Media</b>
